# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Office Manager API
  version: 0.1.0
  description: |
    Web とモバイルクライアントから共通で利用する Office Manager バックエンド API の OpenAPI 仕様です。
    仕様は開発中であり、今後変更される可能性があります。
  contact:
    name: Office Manager Team
    url: https://example.com
    email: dev@example.com
servers:
  - url: https://api.example.com
    description: Production server (replace with actual hostname)
  - url: http://localhost:3000/api
    description: Local development server (Next.js App Router)
...
paths:
  /users:
    get:
      tags: [Users]
      summary: ユーザー一覧の取得
      description: 登録フォームなどで利用するためのユーザー一覧を返します。`officeCode` を指定した場合は該当オフィスのユーザーのみに絞り込みます。
      parameters:
        - name: officeCode
          in: query
          description: フィルタ対象のオフィスコード（大文字・小文字は区別されません）
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserListItem'
        '500':
          $ref: '#/components/responses/UnexpectedError'
    post:
      tags: [Users]
      summary: ユーザー新規登録
      description: フォームデータを受け取り、プロフィール画像をアップロードした上でユーザーを作成します。
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - name
                - email
                - password
                - icon
                - officeCode
              properties:
                name:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
                icon:
                  type: string
                  format: binary
                  description: プロフィール画像ファイル
                officeCode:
                  type: string
                  description: 登録先オフィスのコード（大文字・小文字は区別されません）
      responses:
        '201':
          description: 作成成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserWithOffice'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: メールアドレス重複
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/UnexpectedError'
...
  /users/login:
    post:
      tags: [Users]
      summary: ログイン
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: 認証成功
          content:
            application/json:
              schema:
                type: object
                required:
                  - user
                  - token
                properties:
                  user:
                    $ref: '#/components/schemas/UserSafe'
                  token:
                    type: string
                    description: Bearer トークン（JWT）
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: 認証失敗
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    Office:
      type: object
      required:
        - id
        - code
        - name
      properties:
        id:
          type: integer
          format: int32
        code:
          type: string
        name:
          type: string
        latitude:
          type: number
          format: double
          nullable: true
        longitude:
          type: number
          format: double
          nullable: true
        radiusMeters:
          type: integer
          format: int32
          nullable: true
    UserListItem:
      type: object
      required:
        - id
        - name
        - iconFileName
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        iconFileName:
          type: string
        office:
          allOf:
            - $ref: '#/components/schemas/Office'
          nullable: true
    UserSafe:
      type: object
      required:
        - id
        - name
        - email
        - iconFileName
        - officeId
        - office
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        email:
          type: string
          format: email
        iconFileName:
          type: string
          description: 公開URL
        note:
          type: string
          nullable: true
        entered:
          type: boolean
        enteredAt:
          type: string
          format: date-time
          nullable: true
        exitedAt:
          type: string
          format: date-time
          nullable: true
        officeId:
          type: integer
          format: int32
        office:
          $ref: '#/components/schemas/Office'
...
    UserWithOffice:
      allOf:
        - $ref: '#/components/schemas/UserSafe'
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
        detail:
          type: string
          nullable: true
  responses:
    BadRequest:
      description: リクエストが不正です
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    UnexpectedError:
      description: 予期しないエラーが発生しました
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'